test: test code for commit convention
test code for commit convention test commit message
More detailed explanatory text, if necessary. Wrap it to abount 72 characters or so. In some contexts, the first line is treated as the subject of the commit and the rest the text as the body. The blank line separating the summary from the body is critical (unless you omit the body entirely); various tools like,  and  can get confused if you run the two together.

Explain the problem that this commit is solving. Focus on why you are making this change as opposed to how (the code explains that). Are there side effects or other unintuitive consequeneses of this change? Here's the place to explain them.🎉🎉🎉🎉🎉🎉🎉